backend/pyproject.toml
[tool.poetry]
name = "djangochat"
version = "0.1.0"
description = "djangochat"
authors = ["leonardocloliver <leonardocloliver@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "4.0.4"
djangorestframework = "3.13.1"
Markdown = "3.3.7"
django-filter = "21.1"
gunicorn = "20.1.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pylint = "2.14.1"
pylint-django = "2.5.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
